+++ vertigo-datafactory/src/test/java/io/vertigo/datafactory/search/AbstractSearchManagerTest.java
@@ -18,13 +18,8 @@
 package io.vertigo.datafactory.search;
 
 import java.io.File;
-import java.io.IOException;
-import java.net.URI;
 import java.net.URL;
 import java.net.URLDecoder;
-import java.net.http.HttpClient;
-import java.net.http.HttpRequest;
-import java.net.http.HttpResponse;
 import java.nio.charset.StandardCharsets;
 import java.time.Instant;
 import java.time.LocalDate;
@@ -50,7 +45,6 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import io.vertigo.connectors.elasticsearch.ElasticSearchConnector;
 import io.vertigo.core.lang.VUserException;
 import io.vertigo.core.lang.WrappedException;
 import io.vertigo.core.node.AutoCloseableNode;
@@ -140,39 +134,16 @@ public abstract class AbstractSearchManagerTest {
 		geoCircleFacetDefinition = definitionSpace.resolve("FctLocalisationCircleItem", FacetDefinition.class);
 		geoHashClusterFacetDefinition = definitionSpace.resolve("FctLocalisationHashItem", FacetDefinition.class);
 		itemIndexDefinition = definitionSpace.resolve(indexName, SearchIndexDefinition.class);
-
 		removeAll();
 	}
 
 	@BeforeAll
 	public static void doBeforeClass() throws Exception {
-		//For Embedded ElasticSearch, we must remove data dir in index, in order to support versions updates when testing on PIC
+		//We must remove data dir in index, in order to support versions updates when testing on PIC
 		final URL esDataURL = Thread.currentThread().getContextClassLoader().getResource("io/vertigo/datafactory/search/indexconfig");
 		final File esData = new File(URLDecoder.decode(esDataURL.getFile() + "/data", StandardCharsets.UTF_8.name()));
 		if (esData.exists() && esData.isDirectory()) {
 			recursiveDelete(esData);
-		}
-
-		//for Rest ElasticSearch, we call delete index
-		try {
-			HttpClient client = HttpClient.newHttpClient();
-			deleteIndex("http://localhost:9200/tu_test_idx_item", client);
-			deleteIndex("http://localhost:9200/tu_test_idx_item.metadata", client);
-			deleteIndex("http://localhost:9200/tu_test_idx_item_2", client);
-		} catch (Exception e) {
-			System.err.println("Impossible de contacter Elasticsearch pour supprimer l'index");
-			e.printStackTrace();
-		}
-	}
-
-	private static void deleteIndex(final String ES_URL, HttpClient client) throws IOException, InterruptedException {
-		HttpRequest request = HttpRequest.newBuilder().uri(URI.create(ES_URL)).DELETE()
-				.build();
-		HttpResponse<String> response = client.send(request, HttpResponse.BodyHandlers.ofString());
-		// Code 200 = supprimé avec succès
-		// Code 404 = l'index n'existait pas (ex: premier lancement), on ignore
-		if (response.statusCode() != 200 && response.statusCode() != 404) {
-			System.err.println("Erreur inattendue lors de la suppression de l'index : " + response.body());
 		}
 	}
 
@@ -883,6 +854,17 @@ public abstract class AbstractSearchManagerTest {
 	}
 
 	@Test
+	public void testFacetQueryByStringRange() {
+		index(false);
+		final SearchQuery searchQuery = SearchQuery.builder("QryItemDescriptionRangeFacet")
+				.withCriteria("")
+				.withFacet(EMPTY_SELECTED_FACET_VALUES)
+				.build();
+		var result = doQuery(searchQuery, null);
+		testFacetResultByStringRange(result, "");
+	}
+
+	@Test
 	public void testFacetOptionalFieldByTerm() {
 		testFacetOptionalFieldByTerm("");
 	}
@@ -952,6 +934,38 @@ public abstract class AbstractSearchManagerTest {
 		checkOrderByCount(manufacturerFacet);
 		checkOrderByAlpha(getFacetByName(result, "FctManufacturerItemAlpha"));
 		checkOrderByCount(getFacetByName(result, "FctDescriptionItem"));
+	}
+
+	private void testFacetResultByStringRange(final FacetedQueryResult<Item, SearchQuery> result, final String dbCriteria) {
+		final var dbCount = dbCriteria.isBlank() ? itemDataBase.size() : itemDataBase.containsDescription(dbCriteria);
+		Assertions.assertEquals(dbCount, result.getCount());
+
+		//On vérifie qu'il y a le bon nombre de facettes.
+		Assertions.assertEquals(4, result.getFacets().size());
+
+		final Facet descriptionRangeFacet = getFacetByName(result, "FctDescriptionRangeItem");
+		//On vérifie que l'on est sur le champ Manufacturer
+		Assertions.assertEquals("description", descriptionRangeFacet.getDefinition().getDataField().name());
+		Assertions.assertTrue(descriptionRangeFacet.getDefinition().isRangeFacet());
+
+		//On vérifie qu'il existe une valeur pour peugeot et que le nombre d'occurrences est correct
+		for (final Entry<FacetValue, Long> entry : descriptionRangeFacet.getFacetValues().entrySet()) {
+			final String searchFacetLabel = entry.getKey().label().getDisplay().toLowerCase(Locale.FRENCH);
+			final long searchFacetCount = entry.getValue();
+			if ("#".equals(searchFacetLabel)) {
+				Assertions.assertEquals(7, searchFacetCount);
+			} else if ("a-f".equals(searchFacetLabel)) {
+				Assertions.assertEquals(11, searchFacetCount);
+			} else if ("g-m".equals(searchFacetLabel)) {
+				Assertions.assertEquals(8, searchFacetCount);
+			} else if ("n-s".equals(searchFacetLabel)) {
+				Assertions.assertEquals(11, searchFacetCount);
+			} else if ("t-z".equals(searchFacetLabel)) {
+				Assertions.assertEquals(9, searchFacetCount);
+			} else {
+				Assertions.fail("Unexpected facet " + searchFacetLabel);
+			}
+		}
 	}
 
 	private void testFacetResultByGeo(final FacetedQueryResult<Item, ?> result, final GeoPoint origin) {
@@ -2309,7 +2323,6 @@ public abstract class AbstractSearchManagerTest {
 	}
 
 	private void waitAndExpectIndexation(final long expectedCount, final String queryStr) {
-		searchManager.waitForRefresh(Item.class);
 		final long time = System.currentTimeMillis();
 		long size = -1;
 		try {